pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ViHERMES"
version = "0.1.0"
authors = [
    {name = "Ura Team"}
]
description = "update later"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "chainlit>=1.0.0",
    "neo4j>=6.0.2",
    "openai>=1.0.0",
    "pydantic>=2.7",
    "pydantic-ai>=1.0.15",
    "pydantic-core>=2.16",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.6.1",
    "scikit-learn>=1.4",
    "pydantic-settings>=2.0.0",
    "pymilvus>=2.6.3",
    "python-docx>=1.1.0",
    "typing-extensions==4.12.2",
    "google-generativeai>=0.8.5",
    "tiktoken>=0.12.0",
]



[tool.setuptools.packages.find]
include = ["vihermes*", "data", "lawgraph", "lawrag", "Agents", "config", "preprocess"]

[tool.setuptools]
include-package-data = true


